requirements.txt
gradio
torch
torchvision
pillow
scikit-image
